fix: Correctly process and display data in modes 2, 3, and 4
This commit addresses issues with how modes 2, 3, and 4 process
incoming data from the socket.  The previous implementation of mode2
had several flags to handle edge cases when separating the header
from the content. The implementation was refactored to simplify the
logic by using a buffer and finding the header index "\n\n".
Mode 3 logic was fixed to match the logic changes in mode 2.
Mode 4 was implemented to write the data to a file, truncate the
header and printing the body data.
These changes ensure that the data is displayed correctly, even
when the header and content are split across multiple socket receives.
Add also None validation to avoid TypeErrors